+++ app/views/login.scala.html
@@ -85,7 +85,7 @@
 	<!--
 		NAŠ HELPER
 		-->
-			@helper.form(routes.Application.findUser()) {
+			@helper.form(routes.UserController.findUser()) {
 				<h5>Username:</h5>
 					<input type="text" value="" id="username" name="username" required/>
 					<h6>@usernameMessage</h6>
