+++ app/views/registration.scala.html
@@ -87,7 +87,7 @@
 		
 		MINIMUM of three characters is here!! 
 		-->
-		@helper.form(routes.Application.addUser(),'id->"nameForm") {
+		@helper.form(routes.UserController.addUser(),'id->"nameForm") {
 			<h5>Username:</h5>
 			<input pattern=".{3,}" required title="3 characters minimum" type="text" value="" id="username" name="username" required/>
 			<h6>@usernameMessage</h6>
