+++ app/views/showProduct.scala.html
@@ -1,4 +1,4 @@
-@(usernameSes: String)(name: String)(desc: String)(price: Double)(category: String)
+@(usernameSes: String)(name: String)(desc: String)(price: Double)
 
 @main("Registration") {
 		
@@ -104,11 +104,13 @@
 		
 		<h2><em>Category </em></h2><br>
 		
-			<p>@category</p><br>	
+			<p>Note * ispis String varijable ovdje - Category</p><br>
 		
 		<h2><em>Published by User </em></h2><br>
 		
 			<p>@usernameSes</p><br>	
+			
+			
 			
    
 <!-- FOOTER -->
